requirements.txt
fastapi
tortoise-orm
jinja2
pydantic-settings
python-multipart
uvicorn
gunicorn
celery
redis